Bind transaction data to view holder

+++ app/src/main/java/com/raturu/pertaminanow/data/model/Transaction.kt
@@ -8,4 +8,4 @@ import java.util.*
  * Name       : Zetra
  * GitHub     : https://github.com/zetbaitsu
  */
-data class Transaction(val id: String, val createdAt: Date, val gasolineAmount: Int, val paidAmount: Long)
+data class Transaction(val id: String, val createdAt: Date, val gasolineAmount: Double, val paidAmount: Long)
